defined seeders for users & complaints table

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -18,8 +18,8 @@ class CreateUsersTable extends Migration
             $table->string('username')->unique();
             $table->string('fullname');
             $table->string('email')->unique();
-            $table->integer('mobile')->unsigned()->unique()->nullable();
-            $table->integer('adharno')->unsigned()->unique()->nullable();
+            $table->double('mobile', 10)->unsigned()->unique()->nullable();
+            $table->double('adharno', 12)->unsigned()->unique()->nullable();
             $table->string('password');
             $table->string('role');
             $table->rememberToken();
